requirements.txt
attrs
cattrs
aiohttp
frozendict
